requirements.txt
flask
functions-framework
google-genai
gspread
requests
pydantic
python-dotenv
